refactor: CafeStudyService interface 삭제
